requirements.txt
pillow
tqdm
win32-setctime